// src/models/interfaces.ts
export interface IConnectionProfile extends vscodeMssql.IConnectionInfo {
	profileName: string | undefined;
	savePassword: boolean | undefined;
	emptyPasswordInput: boolean | undefined;
	azureAuthType: AzureAuthType | undefined;
	accountStore: AccountStore | undefined;
	isValidProfile(): boolean | undefined;
	isAzureActiveDirectory(): boolean | undefined;
}